+++ CHANGELOG.md
@@ -11,3 +11,4 @@ Henry
 - Added 'role' field to User schema, changed creating user route so that 'role' will be added to their profile
 - Changed default icon, made small edits to home page
 - Added different landing screens based on role
+- Added schema for marketplace Item, TODO: Add image parameter to item